Move product models into new folder
